Navegación funcionando

+++ SeriesApp/SeriesApp_Entities/Classes/ClsUser.cs
@@ -9,12 +9,10 @@ namespace SeriesApp_Entities.Classes
     public class ClsUser
     {
         #region Public properties
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string UserName { get; set; }
         public string Email { get; set; }
         public string Password { get; set; }
-
-
         #endregion
 
         #region Constructors
